ai-services/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
grpcio-reflection==1.59.0
protobuf==4.24.3

# Deep Learning
# torch/torchvision/torchaudio are installed separately in each Dockerfile
# because ARM64 wheels come from PyPI (not the pytorch whl/cpu index)
transformers==4.34.1
timm==0.9.7

# Computer Vision
# opencv-python-headless: unpinned so pip selects the correct ARM64 wheel
opencv-python-headless
Pillow==10.0.1
scikit-image==0.21.0
# mediapipe removed: no ARM64 Linux wheel exists; use opencv directly for face detection

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
matplotlib==3.8.0

# Signal Processing
numpy==1.24.4
scipy==1.11.3
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.4.2
loguru==0.7.2
prometheus-client==0.17.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0
